+++ src/gui/Blog/Blog.Service/Services/Contracts/ITopicService.cs
@@ -8,5 +8,5 @@ public interface ITopicService
     public Task<List<TopicTableView>> GetUsedTopicsAsync();
     public Task<List<TopicTableView>> GetAllTopicsAsync();
     public Task<TopicTableView> SaveTopicAsync(EntryTopic topic);
-    public Task<bool> DeleteTopicAsync(uint topicId);
+    public Task<bool> DeleteTopicAsync(long topicId);
 }
